+++ packages/core/src/dynamicResources/commands/deleteResource.ts
@@ -7,7 +7,7 @@ import * as vscode from 'vscode'
 import * as nls from 'vscode-nls'
 import { getLogger } from '../../shared/logger/logger'
 import { showConfirmationMessage, showViewLogsMessage } from '../../shared/utilities/messages'
-import { CloudControlClient } from '../../shared/clients/cloudControlClient'
+import { CloudControlClient } from '../../shared/clients/cloudControl'
 import globals from '../../shared/extensionGlobals'
 import { telemetry } from '../../shared/telemetry/telemetry'
 import { millisecondsSince, Result } from '../../shared/telemetry/telemetry'
