+++ packages/core/src/dynamicResources/commands/saveResource.ts
@@ -12,7 +12,7 @@ import { compare, Operation } from 'fast-json-patch'
 import { ResourceNode } from '../explorer/nodes/resourceNode'
 import { ResourceTypeNode } from '../explorer/nodes/resourceTypeNode'
 import { AwsResourceManager } from '../awsResourceManager'
-import { CloudControlClient } from '../../shared/clients/cloudControlClient'
+import { CloudControlClient } from '../../shared/clients/cloudControl'
 import { CloudControl } from 'aws-sdk'
 import globals from '../../shared/extensionGlobals'
 import { telemetry } from '../../shared/telemetry/telemetry'
